Implemented service update and delete functionality

+++ app/src/main/java/com/codeflo/seg2105_final/AdminActivity.java
@@ -52,7 +52,7 @@ public class AdminActivity extends Activity {
             if(task.isSuccessful()){
                 ArrayList<Service> serviceList = new ArrayList<>();
                 for(DocumentSnapshot doc : task.getResult().getDocuments()){
-                    serviceList.add(new Service(doc.getId(), Integer.parseInt((String) doc.get("rate")), null));
+                    serviceList.add(new Service(doc.getId(), Double.parseDouble((String) doc.get("rate")), null));
 
                     ListView listServices = (ListView)findViewById(R.id.serviceList);
                     ServiceAdapter serviceAdapter = new ServiceAdapter(AdminActivity.this, serviceList);
